+++ src/Ext/Techno/Hooks.cpp
@@ -24,37 +24,34 @@ DEFINE_HOOK(0x6F9E50, TechnoClass_AI, 0x5)
 	return 0;
 }
 
-DEFINE_HOOK(5184FD, InfantryClass_TakeDamage_DeathWeaponFired_Check, 8)
+// Issue #237 NotHuman additional animations support
+// Author: Otamaa
+DEFINE_HOOK(0x518505, InfantryClass_TakeDamage_NotHuman, 0x4)
 {
-	GET(InfantryClass*, pThis, ESI);
-	GET(InfantryTypeClass*, pThisType, EAX);
+	GET(InfantryClass* const, pThis, ESI);
 	REF_STACK(args_ReceiveDamage const, ReceiveDamageArgs, STACK_OFFS(0xD0, -0x4));
 
-	if (pThisType->NotHuman)
+	auto Sequenceint = 11;
+	auto const TypeExt = TechnoTypeExt::ExtMap.Find(pThis->GetTechnoType());
+
+	if (TypeExt->NotHumanRandomDeathSequence.Get())
+		Sequenceint = ScenarioClass::Instance->Random.RandomRanged(11, 15);
+
+	if (ReceiveDamageArgs.WH)
 	{
-		auto SeqInt = 11;
-		auto TypeExt = TechnoTypeExt::ExtMap.Find(pThisType);
-
-		if (TypeExt->NHumanUseAllDeathSeq.Get())
-			SeqInt = ScenarioClass::Instance->Random.RandomRanged(11, 15);
-
-		if (ReceiveDamageArgs.WH)
+		if (auto const WarheadExt = WarheadTypeExt::ExtMap.Find(ReceiveDamageArgs.WH))
 		{
-			if (auto const WarheadExt = WarheadTypeExt::ExtMap.Find(ReceiveDamageArgs.WH))
+			auto Warheadsequence = WarheadExt->NotHumanDeathSequence.Get();
+			if (Warheadsequence > 0)
 			{
-				auto WarheadSeq = WarheadExt->InfDeathSequence.Get();
-				if (WarheadSeq > 0)
-				{
-					WarheadSeq = WarheadSeq > 5 ? 5 : WarheadSeq;
-					SeqInt = 10 + WarheadSeq;
-				}
+				Warheadsequence = Math::min(Warheadsequence, 5);
+				Sequenceint = 10 + Warheadsequence;
 			}
 		}
-
-		pThis->PlayAnim(static_cast<Sequence>(SeqInt), true, false);
-
-		return 0x518515;
 	}
 
-	return 0x5185C8;
+	R->ECX(pThis);
+	pThis->PlayAnim(static_cast<Sequence>(Sequenceint), true);
+
+	return 0x518515;
 }