+++ src/Ext/TechnoType/Body.cpp
@@ -106,8 +106,7 @@ void TechnoTypeExt::ExtData::LoadFromINIFile(CCINIClass* const pINI)
 	this->ChronoMinimumDelay.Read(exINI, pSection, "ChronoMinimumDelay");
 	this->ChronoRangeMinimum.Read(exINI, pSection, "ChronoRangeMinimum");
 	this->ChronoDelay.Read(exINI, pSection, "ChronoDelay");
-
-	this->NHumanUseAllDeathSeq.Read(exINI, pSection, "NotHumanUseAllDeathSequence");
+	this->NotHumanRandomDeathSequence.Read(exINI, pSection, "NotHumanRandomDeathSequence");
 
 	// Ares 0.A
 	this->GroupAs.Read(pINI, pSection, "GroupAs");
@@ -149,7 +148,7 @@ void TechnoTypeExt::ExtData::Serialize(T& Stm)
 		.Process(this->ChronoMinimumDelay)
 		.Process(this->ChronoRangeMinimum)
 		.Process(this->ChronoDelay)
-		.Process(this->NHumanUseAllDeathSeq)
+		.Process(this->NotHumanRandomDeathSequence)
 		;
 }
 void TechnoTypeExt::ExtData::LoadFromStream(PhobosStreamReader& Stm)